Usato sql nativo in metodo getGiudiceResponse

+++ src/main/java/it/cgmconsulting/mostracanina_cerullo/response/GiudiceResponse.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,7 +17,7 @@ public class GiudiceResponse {
 
     private String cognome;
 
-    private double avg;
+    private BigDecimal avg;
